updated transformers for compact typography tokens
